server/requirements.txt
fastapi
psycopg2
sqlalchemy-cockroachdb
uvicorn
epicbox
docker
urllib3==1.26.15 